pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "term_extractor"
version = "1.3.0"
description = "Extracts bilingual terminology from a CSV file"
authors = [{ name = "Abe Hasan", email = "a.hasan.contact@gmail.com" }]
readme = "README.md"
dependencies = ["levenshtein==0.27.3",
    "nltk==3.9.2",
    "spacy==3.8.7",
    "sentence-transformers==5.1.2",
    "wordfreq==3.1.1",
    "openpyxl==3.1.5",
    "pywebview>=5.0"]

[tool.setuptools.packages.find]
# With flat layout, packages are in the project root
where = ["."]
include = ["term_extractor*"]

[tool.setuptools.package-data]
# Example: include JSON files inside your package
term_extractor = ["edit-terms.html"]